feat: initial commit
This repo features a new WASI test suite. Each test is a self-contained .wasm
file containing both the code and a custom section containing .toml describing
the setup of the test environment. These .wasm files can be built in GitHub
Actions and distributed as artifacts so that any runtime can use these tests.

Signed-off-by: Nathaniel McCallum <nathaniel@profian.com>